- Refactored some packages to reflect the 3 layers, for better clarity - Refactored APIAction and APIActor to make them simpler to use - Made endpointUrl parameter in RestCall mandatory and added accessors - Added empty APICredentials constructor for better API consistency
