+++ src/main/java/com/jbaxenom/laget/interactions/api/entities/AbstractEntity.java
@@ -1,0 +1,23 @@
+package com.jbaxenom.laget.interactions.api.entities;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+/**
+ * @author jbaxenom on 11/21/15.
+ */
+public abstract class AbstractEntity {
+
+    public String toJSONString() {
+        ObjectMapper mapper = new ObjectMapper();
+        String jsonString = "";
+        try {
+            jsonString = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(this);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+
+        return jsonString;
+    }
+
+}
